Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Are we sure we can't go from "nearly-out-of-power" to "so low that the battery is damaged" in under a minute? 

For that matter, do we even do anything now if we're off and about to damage things? Shouldn't the EC invoke the battery disconnect if it's dangerously low?

Patch-set: 1
Reviewer: Gerrit User 1000058 <1000058@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1
Label: Verified=0