Update patch set 1
Patch Set 1:

As discussed on IM, prepare_tx / start_rx have now a large number of parameters (so we will slowly stacking the last one). It might be better to create a structure for the constant ones ( channel, flags, periph) and just points to it.

Patch-set: 1
Label: Code-Review=0
Label: Verified=0