WIP: Add end-to-end test to check keyboard debouncing
Add a facility in the EC keyboard_scan driver to play back a series of
key scans received in a message from the AP. Each scan is timestamped
with a microsecond time when it becomes live.

Add infrastructure in ectool to send a simple sequence file, which
uses ascii characters, to the EC for playing back. The sequence file
has the expected output from the test, so that the test result can be
determined.

Each new key scan is presented by the EC when its timestamp arrives.
To avoid large numbers, use the concept of 'beats' in the sequence file,
where the length of each beat is configurable. This means that a
sequence with events on beat 0, 1, 2, 3 with a beat length of 20ms will
result in events changing a 0us, 10,000us, 20,000us, 30,000us.

BUG=chrome-os-partner:
TEST=chrome-os-parter:12179
On snow:
./ectool keyscan 10000 key_sequence.txt

See that the test passes.

Change-Id: Id690d6c5181d373c41ce030bd051017324b8ffbf