Update patch set 1
Patch Set 1:

We discussed this.

My impression is that we would get more useful information from collecting a histogram of the delays between successive scans of the same row.  The result of a single run of the proposed test depends only on those delays, and it contains less information than the delay themselves, since it can only tell whether multiple events fell within one of the delays.  

If the delays are uncorrelated (which I believe they are), then we can simulate runs of the test by generating a sequence of random delays with the probability distribution given by the histogram.

Patch-set: 1
Reviewer: Gerrit User 1000028 <1000028@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0
Label: Verified=0