Define key scan parameters and get/set command
At present the keyboard scan parameters are hard-coded, so changing them
requires a new EC image. This can be problematic if we want to adjust
the behavior of keyboard scanning since we must send an EC update.

Define the keyboard scan parameters and commands to get/set these
parameters.

Signed-off-by: Simon Glass <sjg@chromium.org>

BUG=chrome-os-partner:12179
BRANCH=snow,link
TEST=manual
Build for all boards

Change-Id: I715755cb5357503723b27ae33053dba1452e48e0