WIP: Re-factor keyboard scan task
This CL re-factors the keyboard scan task to eliminate reliance on
timers and greatly reduce time spent polling.

Since the polling loop relies only on whether or not inputs are
asserted rather than time constants, we cannot fall out of the loop
prematurely. We also don't spend time polling if no inputs are
asserted.

Any change in keyboard state will cause the host to be notified. The
task stays awake so long as the total number of keys pressed > 0. We
only put the task to sleep after we've notified the host of any
changes and have scanned the matrix to ensure no keys are pressed.

The flow of the task is:
1. If keys have changed, interrupt host
2. If # of keys pressed is zero, unmask interrupts and wait. Otherwise
   skip to 4.
3. When an interrupt is received, mask interrupts.
4. Scan matrix while any key is held until a change is detected.

BRANCH=snow
BUG=chrome-os-partner:12179
Signed-off-by: David Hendricks <dhendrix@chromium.org>
TEST=See notes below

Test methods used:
- Smashing a row of keys several times and comparing length of output
- Holding keys for various lengths to check for repeated keys
- Typing while EC is being re-flashed (i2c saturated)
- Key combinations work, and changes to combinations show up on UI
  near instantly.

Change-Id: I7f481f6b9cc2399d710c05c1fee4e9394f79f1f0