Re-factor flow of keyboard scan task
This CL re-factors the keyboard scan task to eliminate reliance on
timers and greatly reduce time spent polling.

The polling loop relies only on key state. The loop will scan the
matrix first and then break if a change is detected or if keys are
no longer being held (inputs no longer asserted).

The loop will be re-entered if keys were pressed in the previous scan.
If keys are no longer pressed then the debounced state is updated, the
change is reported to the host, and the task put to sleep.

The flow of the task is:
1.  If keys have changed since previous iteration, interrupt host
2a. If # of keys pressed is zero, unmask interrupts and wait. Otherwise
    skip to 3.
2b. When an interrupt is received, mask interrupts.
3.  Scan matrix until a change is detected or no keys are pressed.

BRANCH=snow
BUG=chrome-os-partner:12179
Signed-off-by: David Hendricks <dhendrix@chromium.org>
TEST=See notes below

Test methods used:
- Smashing a row of keys several times and comparing length of output
- Holding keys for various lengths to check for repeated keys
- Typing while EC is being re-flashed and "battery" command looping
- Key combinations work, and changes to combinations show up on UI
  near instantly.

Change-Id: I7f481f6b9cc2399d710c05c1fee4e9394f79f1f0