stm32: fix race condition enabling keyboard interrupt
Previously, if a key was pressed after the last polling but
the interrupts were fully enabled, we were missing the edge.
So dropping the key press event and seeing only the key release.

Now we check if keys are down before waiting for the next interrupt,
if any, we re-start polling immediatly.

The row state reading code is unchanged, just moved to a function
in order to re-use it.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=snow
BUG=chrome-os-partner:7484 chrome-os-partner:12179
TEST=type on the keyboard and do not see any missing key presses.
On instrumented board, record the the matrix scans with a logic analyzer
and stare at the waveforms.

Change-Id: I33a9fb23825626ae5a8482f836482af79702fc17