Add EC keyscan test interface
Add EC commands for managing a list of keyscan events which the EC
should replay instead of its normal key scanning operation.

There are two commands: one adds to the list of events. The other
allows the list to be cleared, the sequence to be started, and the
resulting information to be collected.

BUG=chrome-os-partner:12179
BRANCH=none
TEST=manual for now:
On snow:
./ectool keyscan 10000 key_sequence.txt

See that the test passes.

Signed-off-by: Simon Glass <sjg@chromium.org>

Change-Id: Ie4c3e4d0f5c1dbf642185fec99b9201d47532ae1