Handle arbitration lost on I2C ports
This seems to happen when the I2C signals come up so that the EC sees
a start condition from the remote end.  In this case, the EC refuses
to talk on the I2C port until the EC's I2C state machine is reset.

Also, don't fail on bus-busy, since that's true during a multi-part
transaction such as an I2C string read.

BUG=chrome-os-partner:16262
BRANCH=link

TEST=boot system; 'battery' and 'temps' should give good info
Then run snanda's suspend_stress_test for a while and repeat.

Or a better test is to open 2 crosh shells, sudo bash in each, and
  1) while true; do ectool temps all; sleep 0.5; done
  2) suspend_stress_test

Then watch the EC console for "I2C5 bad status" errors.  These happen
rarely, only on some systems.  With this fix, they'll be reported when
they occur, but should not cause errors to be reported by 'ectool
temps all', since the I2C module will clear the arbitration-lost
status before retrying.

Change-Id: Idfaf9cd7e8ef2abcc0130332890329dd5d2ca052
Signed-off-by: Randall Spangler <rspangler@chromium.org>