link: Disable interrupts while reading/writing bits via onewire
When reading, the line must be sampled in a narrow timing window after
the output pulse.  Interrupts or context switches during this time
corrupt the data.

Similarly, when writing, the difference between a 0-bit and a 1-bit is
the length of the output pulse.  So a context switch or interrupt
there can turn a 1-bit into a 0-bit.

BUG=chrome-os-partner:15507
BRANCH=link
TEST=manual

0. plug in AC power
1. hold down shift key for the duration of this test
2. powerled yellow
3. powerled red
4. repeat steps 2-3 several times
5. release shift key

Power adapter LED should toggle color each time.  (It may also toggle
to the normally expected color during this experiment, if the charging
task updates it.)  Power adapter LED should NOT turn off during this test.

Change-Id: Ief11e6e9a5b07aa3a25c60c50e4e7744a4705713
Signed-off-by: Randall Spangler <rspangler@chromium.org>