Add AP userspace scripts to tweak lightbar colors
We have yet another tweak for the lightbar, but we don't want to update the
EC. This CL adds an init script that runs on the AP at every boot and pokes
the EC to modify the lightbar settings. We have to run it at every boot
because the EC will hibernate after the AP has been off (not suspended) for
an hour on battery power and will lose its settings.

There's a corresponding CL for the ec-utils ebuild that installs the
userspace scripts into the rootfs.

NOTE that in order to cherry-pick this into R23 from ToT, we have to update
ectool as well, because the original R23 ectool doesn't implement the
"lightbar params" command.

However, the EC image that ectool talks to is built from a *different*
branch that already has the ectool update, so this change DOESN'T backport
any of the EC firmware changes that would otherwise be required. Instead, we
just hack a little bit to make the EC firmware compile.

BUG=chrome-os-partner:16827
BRANCH=link
TEST=manual

Build the image for Link, install, reboot.

Run "ectool lightbar params". The output should match what's in
/usr/share/ec/lightbar_params.txt

Original-Change-Id: Ied2304da58d6d27b4f186e621bb187ba9a574e6c
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/41167
Commit-Queue: Bill Richardson <wfrichar@chromium.org>
Reviewed-by: Bill Richardson <wfrichar@chromium.org>
Tested-by: Bill Richardson <wfrichar@chromium.org>
(cherry picked from commit 024f292edf44a9223d6101edc31571df067a8dce)

Change-Id: Iee3f67848c9da92a644083133a70431efabf42aa