stm32mon: skip error when setting UART parity
When using a pty (such as servod ports) to access the EC serial port,
we cannot set the serial port parity.
We need to set it with an external tool (e.g. dut-control),
so we should ignore the error on the stm32mon side

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=none
BUG=none
TEST=Compile and run several times in a row stm32mon on a Goobuntu host,
and see we are no longer getting "Cannot set tty attributes" error.

Change-Id: I7380a5868b4ddcff1c393e3a9ee55967d1c40a7b