Modify stm32mon to better suite autotest use
When used outside chroot on Goobuntu the serial interface fails to
initialize if parity enable bit is set (for a reason not clearly
understood). As a workaround, in case the device is in /dev/pts/*
category do not set the bit, as it will be set as appropriate by
servod later anyways.

In case an operation fails, the exit value should reflect that (so
that the autotest suite using the utility could report failure).

Often when the programming attempt is undertaken soon after reset,
this utility gets overwhelmed with the console output generated by the
EC on resets. Consume the output before proceeding.

BRANCH=None
BUG=chrome-os-partner:15610
TEST=manual
   . used the utility to program Snow EC both inside and outside chroot.

Change-Id: Id25595d35a2a3ca578639cebd508f599e618787c
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>