Modify stm32mon to better suite autotest use
This change introduces the following modifications:

When used outside chroot on Goobuntu the serial interface fails to
initialize if parity enable bit is set (for a reason not clearly
understood). In cases when the UART stream is carried by USB/FTDI
devices this setting could be omitted, so as a workaround in case the
attempt to enable parity fails we try again without enabling it and
hope for the best.

In case an operation fails, the exit value should reflect that (so
that the autotest suite using the utility could report failure).

Often when the programming attempt is undertaken soon after reset,
this utility gets overwhelmed with the console output generated by the
EC on resets. Consume the output before proceeding.

Instead of printing a long set of dots (one per written/read block),
print a spinning wheel instead.

BRANCH=None
BUG=chrome-os-partner:15610
TEST=manual

   . used the utility to program Snow EC both inside and outside
     chroot, it succeeded.

    Observed the failing attempt to set parity when running outside
    chroot.

    Observed spinning characters instead of stream of dots.

Change-Id: Id25595d35a2a3ca578639cebd508f599e618787c
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>