stm32: Dynamically generate GPIO port listing for keyboard outputs
The select_column() function iterates thru a listing of ports and
asserts specified outputs. In the past, we have hard-coded an array
of GPIO ports to iterate thru. This replaces the hard-coded array
with one that is initialized when the keyboard scan task starts.

The downside is that we need a per-board #define to allocate the
exact amount of space needed. The upside is that the code is more
adaptable for new platforms.

Two minor side-effects of this patch:
1. It turns out GPIO_D does not need to be listed in this array for
   current boards, maybe it was used on earlier Daisy/Snow boards.
   Current boards spread keyboard outputs across 2 GPIO ports only.

2. The array needed a more descriptive name so I attempted to give
   it one (kb_out_ports).

BUG=none
BRANCH=none
TEST=Tested on Snow

Signed-off-by: David Hendricks <dhendrix@chromium.org>
Change-Id: Ie97210ff3c99c79f827f7ea49cd1b25ca1d7415d