Update patch set 9
Patch Set 9:

> personally, I don't feel the stack dump in that shape is 
> particularly useful : due to the ABI, the only really useful 
> information in the stack are the return addresses and cutting 
> it to 16 slots will only probably keep between 0 and 2 levels 
> of call.

I don't have a concrete idea but it sounds this feature is 'really handy for debugging' according to the description of the bug.

> if you feel that's useful for you given the code size increase 
> is still acceptable (when PANIC_HELP is not defined)

In patchset #10, I moved the code as much as I can inside '#ifdef CONFIG_PANIC_HELP' so that the size impact will be minimized. Also, if you think dump size is too small, please propose a number on the bug.

The bug also states 'this is particularly true for watchdog panics', which is not addressed by this CL. Currently, wdt interrupt handler uses its own function (watchdog_trace), which dumps a little information. Looking at the code, I think it's better to unify report_panic and watchdog_trace so that both handlers can print more information about the process that was running before the exception.

I will create another CL to touch watchdog code because it can change the behavior of the EC under normal operation.

Patch-set: 9