Update patch set 2
Patch Set 2:

I verified it by implementing usage_fault_handler, which triggers divzero. I found this wasn't clean (because divzero itself is a usage fault). So, I also tried divzero in svc_handler. Before triggering a usage fault in the svc_handler, I pushed a couple of integers to mark the bottom of the stack before exception. So, when I saw the dump, it was pretty clear that the stack was located correctly by the panic handler.

Patch-set: 2
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0