spring: show yellow LED when power is just plugged in
When power is just plugged in, the charging state machine might be in
idle state. This results in LED showing green before changing to yellow.
Let's show yellow for three seconds to let charging state machine
stabilize. For the case of battery fully charged, this looks like the
battery charges for three seconds and then reaches full.

BUG=chrome-os-partner:18914
TEST=Manual on Spring. Unplug the battery and see LED goes red after few
seconds.
BRANCH=spring

Change-Id: Ib158f936d1a3759f055da78376ab9d5dc1748f01
Signed-off-by: Vic Yang <victoryang@chromium.org>