Update patch set 1
Patch Set 1:

Adding DMA is best done in a separate CL.  This one works, and I'd prefer to base future changes on top of this rather than try to keep the whole thing in limbo while trying to figure out why DMA doesn't work.

In the long run we may be able to re-integrate the STM32 variants.

Patch-set: 1