Update patch set 1
Patch Set 1:

A read of zero bytes still puts the address on the bus followed by an immediate stop, right?  That's different than doing nothing.

AKA: a read of zero bytes is the same as a write of zero bytes except that in one case the address includes the low bit set (to indicate read).

From what I could tell i2cdetect in Linux probes some addresses by doing a read of 1 byte and other addresses by doing a write of 0 bytes.  From the docs it indicated that was the safest thing to do at the time of writing based on common/popular devices.

i2cdetect on Linux does not use a read of 0 bytes but since i2c is so open-ended it seems plausible that it could mean something to someone?

Patch-set: 1