Update patch set 1
Patch Set 1:

OK, I take it back--I don't have a solution coded up with the "-1"s.  It will require a bunch more digging on my part to figure out how the stop bit is supposed to be sent with a read of 0 and a lot more testing.

I guess I'll go back to my original question, should we:
* Submit this as is and worry about the rest later?
* Try to fix/test on other i2c implementations (lm4 / stmf)?
* Do something else

This is going to stay on the back burner for me for now.  OK with submitting it as is, someone hijacking, or just leaving it un-submitted and knowing that 'i2cdetect' won't work so well.

Patch-set: 1