spring: Fix a bug that charger task is stuck
When the user powers off the device with power button, the keyscan mutex
gets locked by GAIA power task and is never unlocked until next boot.
This blocks charger task when it tries to send battery key. Let's not
try to do anything with keyscan when the system is off.

BUG=None
TEST=Power off with power button. Plug/unplug charger and see console
output.
BRANCH=spring

Change-Id: Ic6808bd669f054ba48208566de99e837518274fb
Signed-off-by: Vic Yang <victoryang@chromium.org>