slippy: Drive CPU_PGOOD high during power sequencing
I'm not 100% sure about this signal because it seems to have
changed from IvyBridge.

It used to come out of the PCH and go into the CPU, but with
Haswell it seems to expect to be driven externally.

PROCPWRGD from Haswell ULT EDS:
The processor requires this input signal to be a clean
indication that the VCC and VDDQ power supplies are
stable and within specifications. This requirement
applies regardless of the S-state of the processor.
'Clean' implies that the signal will remain low (capable
of sinking leakage current), without glitches, from the
time that the power supplies are turned on until they
come within specification. The signal must then
transition monotonically to a high state.

BUG=chrome-os-partner:19398
BRANCH=none
TEST=manual: Ensure CPU_PGOOD is high when powered on

> gpioget CPU_PGOOD
  1  CPU_PGOOD

Change-Id: I7d46bf2cf5338055eab9c2695ad44d95a8d658a0
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>