Fix detecting inconsistent flash state
If the last bank of flash in a region was protected and the rest of
the region was unprotected, flash_get_protect() shoud return the
INCONSISTENT flag, but this wasn't being properly detected.

Port the existing fix from STM32F.

BUG=chrome-os-partner:19529
BRANCH=none (not likely worth porting to link)
TEST=protect just the last bank of RW firmware, then flashinfo

    > flashinfo
    Physical: 128 KB
    Usable:   128 KB
    Write:     64 B
    Erase:    256 B
    Protect: 4096 B
    Flags:   all_now INCONSISTENT
    Protected now:
        ........ ........ ........ .......Y

Should have the inconsistent flag set.

Change-Id: I407737cef42748da6b3ec40d84968c76ee07972c
Signed-off-by: Randall Spangler <rspangler@chromium.org>