Bus fault handler doesn't need to be naked
When it was a naked call, if it doesn't jump somewhere, it just
falls through to whatever code happens to be next.  Which happened to
re-enable bus faults - meaning that only the first fault was being
ignored.  That's a problem for flash protection on STM32, where writes
to potentially-locked key registers come in pairs.

BUG=chrome-os-partner:19564
BRANCH=spring
TEST=manual, on pit:

  1. Hack in a igbusfault console command to call ignore_bus_fault()
  2. igbusfault 1
  3. ww 0x40023c0c 0xdeadbeef -> no fault
  4. ww 0x40023c0c 0xdeadbeef -> still no fault
  5. igbusfault 0
  6. ww 0x40023c0c 0xdeadbeef -> exception and reboot

Change-Id: I5cf74b06a76ebf0176a2905924be87b3d1c825ce
Signed-off-by: Randall Spangler <rspangler@chromium.org>