slippy/falco: Add back in flash and hiberantion setup code
These were tested to make sure they are not responsible for
bricking boards.

The flash code is necessary for software sync, and it is
essential to be able to update boards without servo.

The hibernate setup code was breaking jumping between images
and the passing of data between the images -- needed for
SW SYNC but also needed for things like "reboot ap-off".

BUG=chrome-os-partner:19366
BRANCH=none
TEST=manual: emerge-slippy chromeos-ec, flash+boot, update again via SW Sync

Change-Id: Ib1363b353f18b21e0cad3209783e1e0b4ff24e86
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>