Update patch set 3
Patch Set 3:

Motivation is really about dealing with upstream.  Upstream really pushed back on i2c arbitration on snow and it took a huge amount of effort to get it in.  It seems like trying to upstream two other "weird" i2c solutions is just going to be even harder.  It would be nice if we could just end up with a single solution.

Patch-set: 3