Update patch set 3
Patch Set 3:

Odds are pretty good.

The big advantage is that the passthru command would be in EC-RO firmware, so we could use it from recovery mode.

Otherwise, even if we later update EC-RW to support the passthru, we'd still need to support the existing commands for the recovery mode kernel.

Patch-set: 3