stm32: avoid getting stuck on interrupted AP I2C transfer
When the main CPU is shutdown, if an AP/EC I2C transfer is still
on-going, the slave I2C controller on the EC will stay stuck with the
"busy" bit set and will refuse further I2C communications when the AP is
restarted.
To solve that issue, we check the busy bit everytime we shut down the
CPU and fully reset the I2C controller if it is set.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=spring
BUG=chromium:241094
TEST=run bootperf on a Spring board for a lot of cycles :
~/trunk/src/platform/crostestutils/bootperf-bin/bootperf 172.22.x.x 40
and see that we are not failing on a ssh connection timeout.

Change-Id: Ia691ae5bfba85ceacd58968487a89913934a3665