Update patch set 4
Patch Set 4:

Thanks for the reminder on the undefined behavior! The reason I chose to implement this with pthread and go through all the complexity is that we get total control of task scheduling this way. With this we can emulate the exact behavior our Cortex RTOS does. The signal handling is for suspending a running thread, which I don't know any other way to do. Perhaps we should switch back to the busy loop to avoid pthread calls in signal handler? Busy loops are evil indeed, but are probably fine since we only run for less than 10 secs and at most one thread is in a busy loop?

Patch-set: 4
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0