Update patch set 1
Patch Set 1:

Thanks for making this completed. LGTM and please feel free to submit.

I have one suggest if you have time: add a test that buf_bytes is NOT times of unit_bytes. For example, the buf1[] size is 5, but the unit_bytes is 2.

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0