Update patch set 2
Patch Set 2:

I should have said something earlier, but I wasn't thinking. HIGH and LOW are really wrong naming conventions when it comes to open drain pins. At least for the EC on the haswell boards.  Setting an open drain pin to 1 enables the FET that ties the input pin to GND.

So what is the definition of HIGH? 1? If so, that is wrong for open drain.

Patch-set: 2